+++ Random.h
@@ -8,9 +8,9 @@
 #endif //ZADANIE_2_RANDOM_H
 
 class Random {
-    uint32_t r;
-    static constexpr uint32_t MULTIPLY_CONST = 279410273;
-    static constexpr uint32_t MODULO_CONST = 4294967291;
+    uint64_t r;
+    static constexpr uint64_t MULTIPLY_CONST = 279410273;
+    static constexpr uint64_t MODULO_CONST = 4294967291;
 public:
     Random(uint32_t seed) {
         r = seed;
